Split selectors into separate files
