+++ front-end/src/selectors/clientSelectors.js
@@ -1,0 +1,12 @@
+import { createSelector } from 'reselect';
+
+const selectClientBody = (state) => state.api.client.body;
+const selectClientConnected = (state) => state.api.client.connected;
+export const selectClient = createSelector(
+  selectClientBody,
+  selectClientConnected,
+  (body, connected) => ({
+    client: body,
+    connected,
+  })
+);
