+++ front-end/src/selectors/formSelectors.js
@@ -1,0 +1,9 @@
+import { createSelector } from 'reselect';
+
+const selectForm = (state) => state.form;
+export const selectFormAttributes = createSelector(selectForm, (form) => ({
+  formData: form.data,
+  error: form.error,
+  submitted: form.submitted,
+  validForm: form.valid,
+}));
