+++ front-end/src/selectors/index.js
@@ -1,34 +1,8 @@
-import { createSelector } from 'reselect';
 export * from './cardSelectors';
 export * from './playerSelectors';
 export * from './turnSelectors';
 export * from './gameStateSelectors';
 export * from './effectSelectors';
-
-export const selectRoundState = (state) => state.round.currentState;
-const selectPlayerReady = (state) => state.round.ready.playerReady;
-const selectTotalReady = (state) => state.round.ready.totalReady;
-export const selectReady = createSelector(
-  selectPlayerReady,
-  selectTotalReady,
-  (playerReady, totalReady) => ({ playerReady, totalReady })
-);
-
-const selectClientBody = (state) => state.api.client.body;
-const selectClientConnected = (state) => state.api.client.connected;
-export const selectClient = createSelector(
-  selectClientBody,
-  selectClientConnected,
-  (body, connected) => ({
-    client: body,
-    connected,
-  })
-);
-
-const selectForm = (state) => state.form;
-export const selectFormAttributes = createSelector(selectForm, (form) => ({
-  formData: form.data,
-  error: form.error,
-  submitted: form.submitted,
-  validForm: form.valid,
-}));
+export * from './roundSelectors';
+export * from './formSelectors';
+export * from './clientSelectors.js';
